requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
azure-keyvault-secrets==4.8.0
azure-identity==1.18.0
azure-cosmos==4.7.0
azure-storage-blob==12.22.0
azure-ai-contentsafety==1.0.0
azure-ai-formrecognizer==3.3.3
openai==1.51.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
reportlab==4.2.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
tenacity==9.0.0
opencensus-ext-azure==1.1.13
